Add files via upload
This is a script I wrote and built. I needed to access Gitlab, find a group, list its projects, get the items, find the .rdl files, parse the information, transform the data within, and then I needed to commit the changes to the same project using the same file path as before. The script creates a branch, commits the changes, and creates a merge request as long as there is no conflicting merges.